services/auth-service/requirements.txt
fastapi
uvicorn
python-dotenv
python-jose[cryptography]
psycopg2-binary
sqlalchemy
